More refactoring, take the static objects of the interfaceBoard namespace.
