+++ v2/current_version.txt
@@ -1,1 +1,1 @@
-2.92
+3.00
