+++ v2/src/shared/Interface.cc
@@ -2,16 +2,22 @@
 #include "Interface.hh"
 #include "InterfaceBoard.hh"
 
-namespace interfaceboard {
 
-InterfaceBoard board;
+// TODO: Make this a proper module.
+#if defined HAS_INTERFACE_BOARD
+
+namespace interface {
+
+
+LiquidCrystal* lcd;
+InterfaceBoard* board;
 
 bool isConnected() {
+
 	// Strategy: Set up the foo pin as an input, turn on pull up resistor,
 	// then measure it. If low, then we probably have an interface board.
 	// If high, we probably don't.
 
-//	MCUCR = MCUCR & ~(1<<4);
 	INTERFACE_FOO_PIN.setValue(true);
 	INTERFACE_FOO_PIN.setDirection(false);
 
@@ -30,31 +36,35 @@ bool isConnected() {
 	}
 
 	return (!INTERFACE_FOO_PIN.getValue());
+
 }
 
-void init() {
-	board.init();
+void init(InterfaceBoard* board_in, LiquidCrystal* lcd_in) {
+    board = board_in;
+    lcd = lcd_in;
 }
 
 void pushScreen(Screen* newScreen) {
-	board.pushScreen(newScreen);
+        board->pushScreen(newScreen);
 }
 
 void popScreen() {
-	board.popScreen();
+        board->popScreen();
 }
 
 void doInterrupt() {
-	board.doInterrupt();
+        board->doInterrupt();
 }
 
 micros_t getUpdateRate() {
-	return board.getUpdateRate();
+        return board->getUpdateRate();
 }
 
 void doUpdate() {
-	board.doUpdate();
+        board->doUpdate();
 }
 
 
 }
+
+#endif
